+++ packages/fighting-design/input/src/input.vue
@@ -81,7 +81,15 @@
   }
 
   /** 样式列表 */
-  const styleList = styles(['placeholderColor', 'background', 'textColor', 'activeBackground', 'activeBorderColor'])
+  const styleList = styles([
+    'placeholderColor',
+    'background',
+    'textColor',
+    'activeBackground',
+    'activeBorderColor',
+    'width',
+    'height'
+  ])
 </script>
 
 <template>